Added toggle option for app
Toggle option pauses the functionality of the app but still leaves it
in the status bar of the Mac. This saves battery as the view is not
constantly updating and reading color values. The toggle app
functionality is toggled by right clicking on the eye dropper icon and
selecting toggle.